Refactor: (BRD-122) OAuth2AuthorizationRequest 생성로직 캡슐화
